Fix visual studio warnings

+++ libSpinWaveGenie/src/Interactions/DM_Z_Interaction.cpp
@@ -63,7 +63,7 @@ void DM_Z_Interaction::calcConstantValues(Cell &cell)
   tmp4 = 0.25 * X * S * sin(phi_r - phi_s);
 
   neighbors.findNeighbors(cell, sl_r, sl_s, min, max);
-  z_rs = neighbors.size();
+  z_rs = static_cast<double>(neighbors.size());
 }
 
 void DM_Z_Interaction::calculateFirstOrderTerms(Cell & /*cell*/, Eigen::VectorXcd & /*elements*/) {}
