+++ libSpinWaveGenie/test/AdaptiveSimpsonTest.cpp
@@ -14,7 +14,7 @@ std::vector<double> constantFunction(std::deque<double> &x)
   result.reserve(x.size());
   for (std::size_t i = 0; i < x.size(); i++)
   {
-    result.push_back(i + 1);
+    result.push_back(static_cast<double>(i) + 1.0);
   }
   return result;
 };
